Reorganize and document ComoposeThemeProvider
