+++ collect_app/src/main/java/org/odk/collect/android/application/ComposeTheme.kt
@@ -8,7 +8,7 @@ import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import org.odk.collect.android.R
-import org.odk.collect.qrcode.mlkit.ComposeThemeProvider
+import org.odk.collect.androidshared.ui.ComposeThemeProvider
 
 /**
  * Changes to this theme should also be made in `theme.xml`
